Prototype v1
This is the prototype which will be used for the demonstration, will be updated to include the final version of the grouping tool when it is available. 

The CSS I was given was using a ready made style sheet from w3schools, so I may try to familiarise myself with this and add it later.

I will be going in to uni tomorrow to attempt to get this live on the server.

"sign-in.php" is your starting point.

C2244667 is the student account, L2244667 is the lecturer account and M2244667 is the moderator account. They all share the same password "testing1!".